stream: implement WritableBase without buffering
